LEDBanners/requirements.txt
acme==1.12.0
Adafruit-Blinka==6.15.0
adafruit-circuitpython-framebuf==1.4.8
adafruit-circuitpython-led-animation==2.5.11
adafruit-circuitpython-neopixel==6.2.3
adafruit-circuitpython-pixel-framebuf==1.1.3
adafruit-circuitpython-pixelbuf==1.1.1
Adafruit-PlatformDetect==3.17.2
Adafruit-PureIO==1.1.9
adafruit-python-shell==1.3.3
appdirs==1.4.4
arandr==0.1.10
args==0.1.0
astroid==2.5.1
asttokens==2.0.4
automationhat==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certbot==1.12.0
certbot-apache==1.10.1
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
clint==0.5.1
colorama==0.4.4
colorzero==1.1
ConfigArgParse==1.2.3
configobj==5.0.6
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
distro==1.5.0
docutils==0.16
dominate==2.6.0
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Colorpicker==0.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fourletterphat==0.1.0
gpiozero==1.6.2
html5lib==1.1
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
josepy==1.2.0
lazy-object-proxy==0.0.0
logilab-common==1.8.1
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mypy==0.812
mypy-extensions==0.4.3
Nuitka==0.6.11.3
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
pantilthat==0.0.7
parsedatetime==2.6
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
psutil==5.8.0
pycairo==1.16.2
pycups==2.0.1
pyftdi==0.53.3
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
PyICU==2.5
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
pynotify==0.1.1
pyOpenSSL==20.0.1
pyRFC3339==1.1
pyserial==3.5
pysmbc==1.0.23
python-apt==2.2.1
python-augeas==0.5.0
pytz==2021.1
pyusb==1.2.1
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
requests-toolbelt==0.9.1
responses==0.12.1
roman==2.0.0
rpi-ws281x==4.3.1
RPi.GPIO==0.7.0
RTIMULib==7.2.1
SCons==4.0.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.2.0
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==2.2.1
spidev==3.5
SQLAlchemy==1.4.28
ssh-import-id==5.10
sysv-ipc==1.1.0
thonny==3.3.14
toml==0.10.1
touchphat==0.0.1
twython==3.8.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
unicornhathd==0.0.4
urllib3==1.26.5
visitor==0.1.3
watchdog==2.1.6
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==3.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.2.0
